Ensure that lockc namespace is privileged
For now it's maybe not so important, but in the future we might want to
enforce lockc policies on the already deployed k8s pods somehow. When
that happens, we need to make sure that the `lockc` k8s namespace
applies privileged policy by default.

Since Helm doesn't provide any way to create namespaces without helmfile
(which might be too much effort for users), we are just providing a
vanilla yaml with the namespace. It's just one command more.

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>